Move libgdf tests to libgdf_cffi directory
