+++ cpp/python/libgdf_cffi/tests/test_hashing.py
@@ -22,7 +22,7 @@ import numpy as np
 from libgdf_cffi import ffi, libgdf
 from librmm_cffi import librmm as rmm
 
-from tests.utils import new_column, unwrap_devary, get_dtype
+from libgdf_cffi.tests.utils import new_column, unwrap_devary, get_dtype
 
 
 @contextmanager
