+++ cpp/python/libgdf_cffi/tests/test_unaryops.py
@@ -8,7 +8,7 @@ import numpy as np
 from libgdf_cffi import ffi, libgdf, GDFError
 from librmm_cffi import librmm as rmm
 
-from tests.utils import new_column, unwrap_devary, get_dtype, gen_rand
+from libgdf_cffi.tests.utils import new_column, unwrap_devary, get_dtype, gen_rand
 
 
 def math_op_test(dtype, ulp, expect_fn, test_fn, nelem=128, scale=1,
