+++ cpp/python/libgdf_cffi/tests/test_validity.py
@@ -6,8 +6,8 @@ import numpy as np
 from libgdf_cffi import libgdf
 from librmm_cffi import librmm as rmm
 
-from tests.utils import new_column, unwrap_devary, get_dtype, gen_rand
-from tests.utils import buffer_as_bits
+from libgdf_cffi.tests.utils import new_column, unwrap_devary, get_dtype, gen_rand
+from libgdf_cffi.tests.utils import buffer_as_bits
 
 
 _dtypes = [np.int32]
